refactor(elysia): rename supersede files, move onError to knowledge/index
- supersede/supersede.ts → list.ts (supersedeListEndpoint)
- supersede/log.ts → create.ts (supersedeCreateEndpoint)
- Move PARSE → 500 onError from knowledge/learn.ts to knowledge/index.ts
  so all three knowledge endpoints share the handler

Matches red PR #882 naming convention. No behavior change.
