+++ src/routes-elysia/knowledge/learn.ts
@@ -1,9 +1,5 @@
 /**
  * POST /api/learn — record a learning pattern.
- *
- * onError PARSE → 500 preserves Hono try/catch semantics (malformed JSON body
- * returned 500 via the old catch block; Elysia defaults to 400 for parse
- * errors, so we remap here to match the HTTP contract tests).
  */
 
 import { Elysia } from 'elysia';
@@ -11,12 +7,6 @@ import { handleLearn } from '../../server/handlers.ts';
 import { LearnBody } from './model.ts';
 
 export const learnEndpoint = new Elysia()
-  .onError(({ code, error, set }) => {
-    if (code === 'PARSE') {
-      set.status = 500;
-      return { error: error instanceof Error ? error.message : 'Parse error' };
-    }
-  })
   .post(
     '/learn',
     ({ body, set }) => {
